+++ plugins/core/tracing/api.go
@@ -23,6 +23,7 @@ import (
 
 var (
 	errParameter = operator.NewError("parameter are nil")
+	errPeerEmpty = operator.NewError("peer is empty")
 )
 
 // CreateEntrySpan creates a new entry span.
@@ -68,8 +69,11 @@ func CreateLocalSpan(operationName string, opts ...SpanOption) (s Span, err erro
 // injector is the injector to inject the context into the carrier.
 // opts is the options to create the span.
 func CreateExitSpan(operationName, peer string, injector Injector, opts ...SpanOption) (s Span, err error) {
-	if operationName == "" || peer == "" || injector == nil {
+	if operationName == "" || injector == nil {
 		return nil, errParameter
+	}
+	if peer == "" {
+		return nil, errPeerEmpty
 	}
 	op := operator.GetOperator()
 	if op == nil {