+++ CHANGES.md
@@ -7,6 +7,7 @@ Release Notes.
 #### Features
 * Add support trace ignore.
 * Enhance the observability of makefile execution.
+* Update the error message if the peer address is empty when creating exit span.
 
 #### Plugins
 * Support [Pulsar](https://github.com/apache/pulsar-client-go) MQ.
@@ -14,6 +15,9 @@ Release Notes.
 * Support http headers collection for Gin.
 * Support higher versions of grpc.
 * Support [go-elasticsearchv8](https://github.com/elastic/go-elasticsearch) database client framework.
+
+### Bug Fixes
+* Fix panic error when root span finished.
 
 0.4.0
 ------------------