+++ CHANGES.md
@@ -15,6 +15,7 @@ Release Notes.
 * Support http headers collection for Gin.
 * Support higher versions of grpc.
 * Support [go-elasticsearchv8](https://github.com/elastic/go-elasticsearch) database client framework.
+* Support `http.Hijacker` interface for mux plugin.
 
 ### Bug Fixes
 * Fix panic error when root span finished.