+++ CHANGES.md
@@ -9,6 +9,7 @@ Release Notes.
 * Enhance the observability of makefile execution.
 * Update the error message if the peer address is empty when creating exit span.
 * Support enhancement go`1.23`.
+* Adding go `1.23` in docker image.
 
 #### Plugins
 * Support [Pulsar](https://github.com/apache/pulsar-client-go) MQ.